third_party/WebKit/LayoutTests/http/tests/security/xss-DENIED-window-name-alert-expected.txt

CONSOLE MESSAGE: line 8: SecurityError: Blocked a frame with origin "null" from accessing a cross-origin frame.
